Refactor process handling: move process-related functions to a dedicated package
