+++ client/ui/process/process_windows.go
@@ -1,4 +1,4 @@
-package main
+package process
 
 import (
 	"os/user"
